refact: ToolBar, EditPanel 관련 모듈 분리 #49
- StickyNoteColorPanel 분리
- getStickyNoteInstance 모듈 분리
  - StickyNote를 얻어오는 함수
- Toolbar 내 도구 별 기능 함수 분리
  - handleSelect, handlePen, handleHand
